requirements.txt
# Azure Functions dependencies
azure-functions
azurefunctions-extensions-bindings-blob

# Azure SDK dependencies
azure-storage-blob
azure-mgmt-web
azure-mgmt-storage
azure-identity
azure-keyvault-secrets
msgraph-sdk==1.5.4

# Data handling and validation dependencies
beautifulsoup4
jsonschema
markdown
markdownify
pillow
pymupdf
python-dotenv
requests
openpyxl==3.1.5
tabulate==0.9.0
azure-search-documents

# AI and NLP dependencies
tiktoken==0.7.0
langchain-text-splitters
openai==1.56.2
cryptography==43.0.3

# Error handling
tenacity==8.5.0

# Web and async dependencies
aiohttp==3.9.0b0
webvtt-py==0.5.1

# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
